+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     - eslint-config-airbnb@8.0.0
     - eslint-plugin-react@5.0.1
 - Fix ```webpack:watch``` (won't stop after the first compilation anymore)
+- Fix npm linking (find global node_modules path correctly)
 
 ## 0.6.3 (2016-04-10)
 - Update dependencies